multiprocessing module is no more used
In previous version of fuzzer.py there was bug connected with deadlock.
If we kill parent, children will endlessly wait.
This can easily be fixed, if we change 'inheritable' mode of pipe's file desctiptor
But it cannot be done straightforwardly by class Pipe's method,
so I decided to use low-level API once again